// api/models/SpaceMemberDeleteRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type SpaceMemberDeleteRequest = {
    /**
     * 空间id
     */
    spaceId: number;
    /**
     * 删除的用户id
     */
    userIdList: Array<number>;
};