// api/models/SpaceMemberAddRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type SpaceMemberAddRequest = {
    /**
     * 空间id
     */
    spaceId: number;
    /**
     * 邀请的用户Id
     */
    userIdList: Array<number>;
};