// api/models/SpaceOwnerTransferRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type SpaceOwnerTransferRequest = {
    /**
     * 空间id
     */
    spaceId: number;
    /**
     * 空间新所有者id
     */
    newOwnerId: number;
};