// api/models/PlayerRevokeRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type PlayerRevokeRequest = {
    /**
     * 空间号
     */
    spaceId: number;
    /**
     * 需要撤销管理的列表
     */
    uidList: Array<number>;
};